requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
dlib==19.24.4
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gradio==4.38.1
gradio_client==1.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
jaconv==0.3.4
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.36.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pykakasi==2.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
sympy==1.13.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
wrapt==1.16.0
